Micrometer library instrumentation
