+++ instrumentation/micrometer/micrometer-1.5/library/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+  id("otel.library-instrumentation")
+}
+
+dependencies {
+  library("io.micrometer:micrometer-core:1.5.0")
+
+  testImplementation(project(":instrumentation:micrometer:micrometer-1.5:testing"))
+}
