+++ instrumentation/micrometer/micrometer-1.5/testing/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+  id("otel.java-conventions")
+}
+
+dependencies {
+  api(project(":testing-common"))
+
+  api("io.micrometer:micrometer-core:1.5.0")
+}
