+++ settings.gradle.kts
@@ -283,6 +283,8 @@ include(":instrumentation:logback:logback-mdc-1.0:library")
 include(":instrumentation:logback:logback-mdc-1.0:testing")
 include(":instrumentation:methods:javaagent")
 include(":instrumentation:micrometer:micrometer-1.5:javaagent")
+include(":instrumentation:micrometer:micrometer-1.5:library")
+include(":instrumentation:micrometer:micrometer-1.5:testing")
 include(":instrumentation:mongo:mongo-3.1:javaagent")
 include(":instrumentation:mongo:mongo-3.1:library")
 include(":instrumentation:mongo:mongo-3.1:testing")
